feat: add world_model module and move relative_map into it
