+++ modules/world_model/relative_map/dag/relative_map.dag
@@ -1,0 +1,11 @@
+module_config {
+    module_library : "/apollo/bazel-bin/modules/world_model/relative_map/librelative_map_component.so"
+    timer_components {
+        class_name : "RelativeMapComponent"
+        config {
+            name: "relative_map"
+            flag_file_path:  "/apollo/modules/world_model/relative_map/conf/relative_map.conf"
+            interval: 100
+      }
+   }
+}
