+++ modules/world_model/relative_map/navigation_lane.h
@@ -28,11 +28,12 @@
 #include <utility>
 
 #include "modules/common/vehicle_state/proto/vehicle_state.pb.h"
-#include "modules/common/vehicle_state/vehicle_state_provider.h"
 #include "modules/common_msgs/localization_msgs/localization.pb.h"
-#include "modules/common_msgs/planning_msgs/navigation.pb.h"
-#include "modules/map/relative_map/proto/relative_map_config.pb.h"
 #include "modules/common_msgs/perception_msgs/perception_obstacle.pb.h"
+#include "modules/common_msgs/planning_msgs/navigation.pb.h"
+#include "modules/world_model/relative_map/proto/relative_map_config.pb.h"
+
+#include "modules/common/vehicle_state/vehicle_state_provider.h"
 
 /**
  * @namespace apollo::relative_map
