+++ modules/world_model/relative_map/navigation_lane_test.cc
@@ -19,16 +19,19 @@
  * @brief This file provides several unit tests for the class "NavigationLane".
  */
 
-#include "modules/map/relative_map/navigation_lane.h"
+#include "modules/world_model/relative_map/navigation_lane.h"
+
+#include "gtest/gtest.h"
+
+#include "nlohmann/json.hpp"
+
+#include "modules/common_msgs/chassis_msgs/chassis.pb.h"
+#include "modules/common_msgs/planning_msgs/navigation.pb.h"
+#include "modules/world_model/relative_map/proto/relative_map_config.pb.h"
 
 #include "cyber/common/file.h"
-#include "gtest/gtest.h"
-#include "modules/common_msgs/chassis_msgs/chassis.pb.h"
 #include "modules/common/vehicle_state/vehicle_state_provider.h"
-#include "modules/map/relative_map/common/relative_map_gflags.h"
-#include "modules/common_msgs/planning_msgs/navigation.pb.h"
-#include "modules/map/relative_map/proto/relative_map_config.pb.h"
-#include "nlohmann/json.hpp"
+#include "modules/world_model/relative_map/common/relative_map_gflags.h"
 
 namespace apollo {
 namespace relative_map {
