+++ modules/world_model/relative_map/proto/BUILD
@@ -3,6 +3,7 @@ load("@com_google_protobuf//bazel:proto_library.bzl", "proto_library")
 load("@com_google_protobuf//bazel:cc_proto_library.bzl", "cc_proto_library")
 load("@com_google_protobuf//bazel:py_proto_library.bzl", "py_proto_library")
 
+
 package(default_visibility = ["//visibility:public"])
 
 cc_proto_library(
@@ -21,7 +22,6 @@ py_proto_library(
     name = "relative_map_config_py_pb2",
     deps = [":relative_map_config_proto"],
 )
-
 cc_proto_library(
     name = "navigator_config_cc_proto",
     deps = [
