+++ modules/world_model/relative_map/relative_map.cc
@@ -14,13 +14,13 @@
  * limitations under the License.
  *****************************************************************************/
 
-#include "modules/map/relative_map/relative_map.h"
+#include "modules/world_model/relative_map/relative_map.h"
 
 #include "cyber/common/file.h"
 #include "modules/common/math/vec2d.h"
 #include "modules/common/util/util.h"
 #include "modules/common/vehicle_state/vehicle_state_provider.h"
-#include "modules/map/relative_map/common/relative_map_gflags.h"
+#include "modules/world_model/relative_map/common/relative_map_gflags.h"
 
 namespace apollo {
 namespace relative_map {
