+++ modules/world_model/relative_map/testdata/multi_lane_map/localization_info.pb.txt
