+++ modules/world_model/relative_map/testdata/multi_lane_map/middle.smoothed
