+++ modules/world_model/relative_map/tools/BUILD
@@ -24,7 +24,7 @@ cc_binary(
         "//cyber",
         "//modules/common/adapters:adapter_gflags",
         "//modules/common/util",
-        "//modules/map/relative_map:relative_map_lib",
+        "//modules/world_model/relative_map:relative_map_lib",
     ],
 )
 
@@ -36,8 +36,8 @@ cc_binary(
         "//cyber",
         "//modules/common/adapters:adapter_gflags",
         "//modules/common/util",
-        "//modules/map/relative_map:relative_map_lib",
-        "//modules/map/relative_map/proto:navigator_config_cc_proto",
+        "//modules/world_model/relative_map:relative_map_lib",
+        "//modules/world_model/relative_map/proto:navigator_config_cc_proto",
         "@com_github_nlohmann_json//:json",
     ],
 )
